"修复了部分bug，1:从微信获取openid,2:更新了oss，3:修复了部分信息获取错误，4:更新数据类型bigint"

+++ transaction/src/main/java/com/hust/trade/transaction/controller/AddCommentController.java
@@ -28,13 +28,13 @@ public class AddCommentController {
    *  增加评论
    * @param userId 用户唯一标识
    * @param userComment 用户具体的评论
-   * @param messageId
-   * @param messageUserId
-   * @return
+   * @param messageId 消息id
+   * @param messageUserId 消息的发布者user_id
+   * @return IsComment对象
    */
   @Transactional
   @PostMapping("/addComment/{userId}/{messageId}/{messageUserId}")
-  public IsComment addComment(@PathVariable Integer userId, @RequestBody String userComment, @PathVariable Integer messageId, @PathVariable Integer messageUserId) {
+  public IsComment addComment(@PathVariable Long userId, @RequestBody String userComment, @PathVariable Long messageId, @PathVariable Long messageUserId) {
     return new IsComment().isTrue(userId, messageId, userComment, messageUserId, userService, commentService, newMessageService, attendService);
   }
